backend/api-gateway/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy[asyncio]==2.0.27
aiosqlite==0.20.0
asyncpg==0.29.0
pydantic[email]==2.6.1
pydantic-settings==2.2.1
python-multipart==0.0.9
httpx==0.26.0
structlog==24.1.0
cryptography==42.0.5
redis[hiredis]==5.0.1
reportlab==4.1.0
prometheus-fastapi-instrumentator==7.0.0
